adding call to insert single nonprofit row to db
